Improved stability of library
Removed processReadyStateChange() helper method due to a scope-related
issue where if the method was invoked by any means other than VIA an
XMLHttpRequest object's event handler, then "this" in the helper method
referenced the incorrect object and broke the functionality.

This update means there is some duplicate code, but results in improved
stability for the aforementioned reason, as well as a less polluted
scope of the SimpleAjax object.